remove duplicate AddPlantForm and update references
Deleted unnecessary AddPlantForm component from components/ and replaced it with the one from pages/. Updated CropCard file path to point to the correct AddPlantForm location
